pyproject.toml
[project]
name = "soundchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "asyncpg>=0.31.0",
    "clickhouse-connect>=0.10.0",
    "django>=6.0",
    "fastapi>=0.126.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "gunicorn>=23.0.0",
    "orjson>=3.11.5",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "python-decouple>=3.8",
    "redis>=7.1.0",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.39.0",
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.2.8",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.11.1",
    "ruff>=0.14.10",
    "types-redis>=4.6.0.20241004",
]
